版本更新-代码编写

+++ enn-energy-price-web/src/main/java/com/enn/energy/price/web/controller/proxyelectricityprice/ProxyElectricityPriceManagerController.java
@@ -102,7 +102,20 @@ public class ProxyElectricityPriceManagerController {
     @ApiOperation( "修改电价版本" )
     public RdfaResult<Boolean> updatePriceVersion(@RequestBody @Valid ElectricityPriceVersionUpdateReqVO electricityPriceVersionUpdateReqVO){
         ElectricityPriceVersionUpdateBO electricityPriceVersionUpdateBO = ElectricityPriceVersionUpdateConverMapper.INSTANCE.electricityPriceVersionUpdateReqVOToBO( electricityPriceVersionUpdateReqVO );
-        return proxyElectricityPriceManagerService.updatePriceVersion( electricityPriceVersionUpdateBO );
+        String lockKey =  String.format("%s:%s:%s", CommonConstant.RedisKey.LOCK_PROXY_PRICE_VERSION_UPDATE_PREFIX,
+                tenantId, electricityPriceVersionUpdateBO.getId());
+        Lock lock = null;
+        try {
+            lock = redDisLock.lock(lockKey  );
+            if (ObjectUtil.isNull( lock )) {
+                return RdfaResult.fail( ErrorCodeEnum.RETRY_AFTER.getErrorCode(), ErrorCodeEnum.RETRY_AFTER.getErrorMsg() );
+            }
+            return proxyElectricityPriceManagerService.updatePriceVersion( electricityPriceVersionUpdateBO );
+        } catch (LockFailException e) {
+            return RdfaResult.fail(ErrorCodeEnum.REIDS_LOCK_ERROR.getErrorCode(), ErrorCodeEnum.REIDS_LOCK_ERROR.getErrorMsg());
+        }finally {
+            redDisLock.unlock(lockKey);
+        }
     }
 
     @PostMapping("/deletePriceVersion")
@@ -112,7 +125,20 @@ public class ProxyElectricityPriceManagerController {
             return RdfaResult.fail( ErrorCodeEnum.VERSION_IS_NOT_ALLOW_DELETE.getErrorCode(), ErrorCodeEnum.VERSION_IS_NOT_ALLOW_DELETE.getErrorMsg());
         }
         ElectricityPriceVersionDeleteBO electricityPriceVersionDeleteBO = ElectricityPriceVersionUpdateConverMapper.INSTANCE.electricityPriceVersionDeleteReqVOToBO( electricityPriceVersionDeleteReqVO );
-        return proxyElectricityPriceManagerService.deletePriceVersion(electricityPriceVersionDeleteBO);
+        String lockKey =  String.format("%s:%s:%s", CommonConstant.RedisKey.LOCK_PROXY_PRICE_VERSION_UPDATE_PREFIX,
+                tenantId, electricityPriceVersionDeleteBO.getId());
+        Lock lock = null;
+        try {
+            lock = redDisLock.lock( lockKey );
+            if (ObjectUtil.isNull( lock )) {
+                return RdfaResult.fail( ErrorCodeEnum.RETRY_AFTER.getErrorCode(), ErrorCodeEnum.RETRY_AFTER.getErrorMsg() );
+            }
+            return proxyElectricityPriceManagerService.deletePriceVersion(electricityPriceVersionDeleteBO);
+        } catch (LockFailException e) {
+            return RdfaResult.fail(ErrorCodeEnum.REIDS_LOCK_ERROR.getErrorCode(), ErrorCodeEnum.REIDS_LOCK_ERROR.getErrorMsg());
+        }finally {
+            redDisLock.unlock(lockKey);
+        }
     }
 
     /**
